Implemented timer using pthread
This patch adds a live timer that shows how much time has elapsed since
starting the program.

The timer is implemented using a separate thread created with
pthread_create().

Additionally the input function is also separated into another thread.

In order to not break the ncurses interface the threads have to do a
pthread_cond_wait() so that they dont try to write on the screen at the
same time... Unfortunately this adds a lot to the cpu-usage. This is an
issue to be fixed later.

this patch also adds a header file "ctyping.h" and renames the .c file
to "ctyping.c".

For now Makefile is changed to use clang but that issue will be adressed soon.